+++ netsuite/__main__.py
@@ -1,3 +1,5 @@
+import json
+import sys
 import logging
 import logging.config
 
@@ -21,7 +23,6 @@ def _set_log_level(log_level):
 @argh.arg('-l', '--log-level', help='The log level to use')
 @argh.arg('-p', '--config-path', help='The config file to get settings from')
 @argh.arg('-c', '--config-section', help='The config section to get settings from')
-@argh.arg('-p', '--config-path')
 def interact(
     log_level=None,
     config_path=DEFAULT_INI_PATH,
@@ -41,16 +42,54 @@ Available vars:
     `ns` - NetSuite client
 
 Example usage:
-    results = ns.getList('customer', internalIds=[1337])
+    ws_results = ns.getList('customer', internalIds=[1337])
+    restlet_results = ns.restlet.request(987)
 """
 
     IPython.embed(
         user_ns=user_ns,
         banner1=banner1,
-        config=traitlets.config.Config(colors='LightBG')
+        config=traitlets.config.Config(colors='LightBG'),
+        # To fix no colored input we pass in `using=False`
+        # See: https://github.com/ipython/ipython/issues/11523
+        # TODO: Remove once this is fixed upstream
+        using=False,
     )
 
 
+@argh.arg('-l', '--log-level', help='The log level to use')
+@argh.arg('-p', '--config-path', help='The config file to get settings from')
+@argh.arg('-c', '--config-section', help='The config section to get settings from')
+def restlet(
+    script_id,
+    payload,
+    deploy=1,
+    log_level=None,
+    config_path=DEFAULT_INI_PATH,
+    config_section=DEFAULT_INI_SECTION
+):
+    """Make requests to restlets"""
+
+    _set_log_level(log_level)
+    conf = config.from_ini(path=config_path, section=config_section)
+    ns = netsuite.NetSuite(conf)
+
+    if not payload:
+        payload = None
+    elif payload == '-':
+        payload = json.load(sys.stdin)
+    else:
+        payload = json.loads(payload)
+
+    resp = ns.restlet.raw_request(
+        script_id=script_id,
+        deploy=deploy,
+        payload=payload,
+        raise_on_bad_status=False,
+    )
+    return resp.text
+
+
 command_parser = argh.ArghParser()
-command_parser.add_commands([interact])
+command_parser.add_commands([interact, restlet])
 main = command_parser.dispatch